products refactoring
